+++ lib/services.dart
@@ -9,3 +9,5 @@ export 'src/services/scheduler/overlap_behaviour.dart';
 export 'src/services/scheduler/schedule.dart';
 export 'src/services/scheduler/scheduler_service.dart';
 export 'src/services/scheduler/scheduled_task.dart';
+
+export 'src/services/abstractions/config_service.dart';